Add files via upload
This is an Excel file for monitoring the intensity of a tropical cyclone. Users should first enter the position for the typhoon every 6 hours and do a Dvorak analysis every time. There will be functions to get the estimated wind speed and calculate its speed and distance to Hong Kong.